So What Did I Chnage?
The problems we have so far is that the "'" is not showing somehow in that method bruv its jhi weird. The next problem is making all the class functions iterate and be called idk why I forgot how to call the class functions. Then the countdown method needs to be fixed and the guesses idea needs to be fleshed out becasue right now we have no way of getting user input or making a user interface to hold the guesses.